% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_HSV_color_by_mask.R
\name{apply_HSV_color_by_mask}
\alias{apply_HSV_color_by_mask}
\title{apply color to pixel.array by mask.}
\usage{
apply_HSV_color_by_mask(pixel.array, pixel.idx, target.color, mask_extreme)
}
\arguments{
\item{pixel.array}{image array as loaded via duflor::load_image()}

\item{pixel.idx}{list declaring pixels to which target.color.hsv is applied}

\item{target.color}{color in rgb-format \code{0-255}, or a member of  \code{colors()}}

\item{mask_extreme}{by default, only the \code{hue}- and \code{saturation}-components
are applied to the pixels. Set this argument to \code{TRUE} to also apply \code{value}.
This will increase the contrast of the image drastically, but might result in
less favorable images.)}
}
\value{
\code{pixel.array} with hsv-values of pixels at positions \code{pixel.idx} modified.
}
\description{
If \code{target.color} is a character-string (e.g. \code{"red1"}) are members of \code{colors()},
it will be translated into its HSV-representation.
If \code{target.color} is a vector of RGB-values, it gets translated into its
HSV-representation.
The resulting HSV-color will be applied to all elements of \code{pixel.array}
which are members of \code{pixel.idx} are translated to their HSV-representation.
}
\details{
Values for \code{target.color} which lie outside the RGB-range of \verb{[0-255,0-255,0-255]} will be
constricted to limits of the range \verb{[0,255]}
}
