% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_hsv_color_to_image_subset.R
\name{apply_hsv_color_to_image_subset}
\alias{apply_hsv_color_to_image_subset}
\title{apply hsv-color to index-located pixels in array}
\usage{
apply_hsv_color_to_image_subset(
  pixel.array,
  pixel.idx,
  target.color.hsv,
  mask_extreme = FALSE
)
}
\arguments{
\item{pixel.array}{image array as loaded via duflor::load_image()}

\item{pixel.idx}{list declaring pixels to which target.color.hsv is applied}

\item{target.color.hsv}{hsv-formatted color to apply}

\item{mask_extreme}{by default, only the \code{hue}- and \code{saturation}-components
are applied to the pixels. Set this argument to \code{TRUE} to also apply \code{value}.
This will increase the contrast of the image drastically, but might result in
less favorable images.)}
}
\value{
modified pixel.array
}
\description{
apply hsv-color to index-located pixels in array
}
\keyword{internal}
