% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_pixel_color_HSV.R
\name{change_pixel_color_HSV}
\alias{change_pixel_color_HSV}
\title{handles color-translation from character-name to hsv and returns results of \code{duflor::apply_hsv_color_to_image_subset()}}
\usage{
change_pixel_color_HSV(pixel.array, pixel.idx, target.color, mask_extreme)
}
\arguments{
\item{pixel.array}{image array as loaded via duflor::load_image()}

\item{pixel.idx}{list declaring pixels to which target.color.hsv is applied}

\item{target.color}{color in rgb-format \code{0-255}, or a member of  \code{colors()}}

\item{mask_extreme}{by default, only the \code{hue}- and \code{saturation}-components
are applied to the pixels. Set this argument to \code{TRUE} to also apply \code{value}.
This will increase the contrast of the image drastically, but might result in
less favorable images.)}
}
\value{
\code{pixel.array} with hsv-values of pixels at positions \code{pixel.idx} modified.
}
\description{
handles color-translation from character-name to hsv and returns results of \code{duflor::apply_hsv_color_to_image_subset()}
}
