% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{check_javaVM_setup}
\alias{check_javaVM_setup}
\title{Determine whether or not the java-VM required for "RBioFormats" is set up}
\usage{
check_javaVM_setup()
}
\value{
boolean value whether or not the JAVA-VM is set up or not
}
\description{
Determine whether or not the java-VM required for "RBioFormats" is set up
}
\keyword{internal}
