% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{diagonal_adjacency}
\alias{diagonal_adjacency}
\title{add 8D-adjacency-grouping to \code{pixel.idx}-object}
\usage{
diagonal_adjacency(pixel.idx)
}
\arguments{
\item{pixel.idx}{list declaring pixels to which target.color.hsv is applied}
}
\value{
\code{pixel.idx} with added 3rd column \code{clus} mapping to a cluster
}
\description{
The function assigns clusters to all coordinate-pairs in \code{pixel.idx}.
A cluster contains all pixels which share a diagonal link with each other.
}
\details{
This means that points \code{(1/1)}, \code{(1/2)} and \code{(2/2)} are assigned the same cluster.
Additionally, points  \code{(5/5)} and \code{(6/6)} are assigned the same cluster.

To consider diagonal matches as well, see \code{\link[=adjacency]{adjacency()}}

Reference: \url{https://stackoverflow.com/a/37946855}
}
