% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.main_args}
\alias{.main_args}
\title{pseudo-function to define default parameter documentation}
\usage{
.main_args(
  pixel.array = NA,
  indicator.array = NA,
  pixel.idx = NA,
  target.color = NA,
  target.color.hsv = NA,
  image.path = NA,
  plot_indicator = NA,
  get_indicator = NA,
  mask_extreme = NA,
  fast_eval = NA,
  return_hsv = NA,
  subset_only = NA,
  upper_bound = NA,
  lower_bound = NA,
  crop_left = NA,
  crop_right = NA,
  crop_top = NA,
  crop_bottom = NA
)
}
\arguments{
\item{pixel.array}{image array as loaded via duflor::load_image()}

\item{indicator.array}{\code{pixel.array} as returned}

\item{pixel.idx}{list declaring pixels to which target.color.hsv is applied}

\item{target.color}{color in rgb-format \code{0-255}, or a member of  \code{colors()}}

\item{target.color.hsv}{hsv-formatted color to apply}

\item{image.path}{path to image-file}

\item{plot_indicator}{control whether or not the indicator image should be plotted}

\item{get_indicator}{control whether or not target pixels should get colored in the indicator-images}

\item{mask_extreme}{by default, only the \code{hue}- and \code{saturation}-components
are applied to the pixels. Set this argument to \code{TRUE} to also apply \code{value}.
This will increase the contrast of the image drastically, but might result in
less favorable images.)}

\item{fast_eval}{Use \verb{C++}-code where possible to reduce execution time?}

\item{return_hsv}{TRUE by default. Controls whether or not function returns pixel-data in \code{HSV}-colorspace or in \code{RGB}}

\item{subset_only}{do you want to load only a subset range of values}

\item{upper_bound}{EITHER:
\itemize{
\item list of upper HSV-bounds, e.g. \code{list(green = c(H_green_lower,S_green_lower,V_green_lower),drought = c(H_drought_lower,S_drought_lower,V_drought_lower))}
\item single vector of length 3 declaring a set of HSV-values
}}

\item{lower_bound}{see \code{upper_bound}}

\item{crop_left}{number of pixels to crop from the left edge of the image}

\item{crop_right}{see \code{crop_left}}

\item{crop_top}{see \code{crop_left}}

\item{crop_bottom}{see \code{crop_left}}
}
\value{
nothing. This function does \emph{literally nothing at all}
}
\description{
pseudo-function to define default parameter documentation
}
\keyword{internal}
