% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{onLoad-routine: Initialises package configuration}
\usage{
.onLoad(lib, pkg)
}
\arguments{
\item{lib}{/}

\item{pkg}{/}
}
\description{
The function initialises various options used by the package upon loading:
\enumerate{
\item The following options are initialised:
\itemize{
\item \code{duflor.default_hsv_spectrums}
\item \code{duflor.default_identifier_area}
}
}
}
\note{
When a package \code{pkg} is attached (see Link below) via \code{library(pkg)}, or the first time
a function of the package \code{pkg} is called via \code{pkg::fun()}, this function
gets called.
}
\seealso{
The differences between \code{attaching} and \code{loading} a package are somewhat
difficult to understand. I found these explanations on
\itemize{
\item \href{https://r-pkgs.org/dependencies-mindset-background.html#sec-dependencies-attach-vs-load}{r-pkgs.org}
\item and \href{https://forum.posit.co/t/when-to-use-onload-vs-onattach/21953/2}{forum.posit.co}
useful.
}
}
\keyword{internal}
