% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onUnLoad}
\alias{.onUnLoad}
\title{onUnLoad-routine: Tears down package configuration}
\usage{
.onUnLoad(lib, pkg)
}
\arguments{
\item{lib}{/}

\item{pkg}{/}
}
\description{
The function unsets various options used by the package:
}
\details{
\itemize{
\item "duflor.default_hsv_spectrums"
\item "duflor.default_identifier_area"
}
}
\keyword{internal}
