% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duflor-package.R
\docType{package}
\name{duflor-package}
\alias{duflor}
\alias{duflor-package}
\title{duflor: Plant Image Analysis For Determination of Leaf- and Root-Area}
\description{
Determine 2-dimensional plant area via known area of an identifier dot.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Claudius-Appel/duflor/}
  \item Report bugs at \url{https://github.com/Claudius-Appel/duflor/issues}
}

}
\author{
\strong{Maintainer}: Claudius Appel \email{151634114+Claudius-Appel@users.noreply.github.com}

}
\keyword{internal}
