% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pixels_HSV.R
\name{extract_pixels_HSV}
\alias{extract_pixels_HSV}
\title{Title}
\usage{
extract_pixels_HSV(
  pixel.array,
  lower_bound,
  upper_bound,
  target.color = "hotpink",
  get_indicator = FALSE,
  plot_indicator = FALSE,
  mask_extreme = FALSE,
  return_hsv = TRUE
)
}
\arguments{
\item{pixel.array}{image array as loaded via duflor::load_image()}

\item{lower_bound}{see \code{upper_bound}}

\item{upper_bound}{EITHER:
\itemize{
\item list of upper HSV-bounds, e.g. list(green = c(H_green_lower,S_green_lower,V_green_lower),drought = c(H_drought_lower,S_drought_lower,V_drought_lower))
\item single vector of length 3 declaring a set of HSV-values
}}

\item{target.color}{color in rgb-format \code{0-255}, or a member of  \code{colors()}}

\item{get_indicator}{control whether or not target pixels should get colored in the indicator-images}

\item{plot_indicator}{control whether or not the indicator image should be plotted}

\item{mask_extreme}{by default, only the \code{hue}- and \code{saturation}-components
are applied to the pixels. Set this argument to \code{TRUE} to also apply \code{value}.
This will increase the contrast of the image drastically, but might result in
less favorable images.)}

\item{return_hsv}{TRUE by default. Controls whether or not function returns pixel-data in \code{HSV}-colorspace or in \code{RGB}}
}
\value{
EITHER:
list-object with the following elements (when supplying one one pair of bounds)
\itemize{
\item \code{pixel.idx} - pixel-locations of pixels detected between lower and upper bound.
\item \code{pixel.count} - number of pixels detected between lower and upper bound
\item \code{img.fraction} - fraction of image detected between lower and upper bound
\item \code{original.img} - fed-in pixel.array
or
list-object with the following elements (when supplying one one pair of bounds)
\item \code{pixel.idx} - pixel-locations of pixels detected between lower and upper bound.
\item \code{pixel.count} - number of pixels detected between lower and upper bound
\item \code{img.fraction} - fraction of image detected between lower and upper bound
\item \code{original.img} - fed-in pixel.array
\item \code{indicator.img} - \code{pixel.array} with hsv-values of pixels at positions \code{pixel.idx} modified.
}
}
\description{
Title
}
