% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{func_old}
\alias{func_old}
\title{a dummy function}
\usage{
func_old(x, y)
}
\arguments{
\item{x}{\itemize{
\item numeric
}}

\item{y}{\itemize{
\item numeric
}}
}
\value{
sum of (x,y)
}
\description{
with even more dummy documentation
}
\examples{
func_old(1,2)
}
