% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_area_vector}
\alias{generate_area_vector}
\title{Generate Random Area Values}
\usage{
generate_area_vector(n, min_area = 100, max_area = 10000, na_prob = 0.2)
}
\arguments{
\item{n}{Anzahl der Werte.}

\item{min_area}{Minimale Flaeche.}

\item{max_area}{Maximale Flaeche.}

\item{na_prob}{Wahrscheinlichkeit fuer fehlende Werte.}
}
\value{
Numeric vector with area values.
}
\description{
Erzeugt einen numerischen Vektor mit zufaelligen Flaechenwerten.
}
