% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_localities.R
\name{generate_localities}
\alias{generate_localities}
\title{Generiert synthetische Lokalitaetsdaten}
\usage{
generate_localities(n_lokalitaeten = 20, seed = 123)
}
\arguments{
\item{n_lokalitaeten}{Integer. Anzahl der zu generierenden Lokalitaeten (max. 20). Default ist 20.}

\item{seed}{Integer. Seed fuer den Zufallszahlengenerator. Default ist 123.}
}
\value{
Ein tibble mit den generierten Lokalitaetsdaten.
}
\description{
Diese Funktion erstellt eine Lookup-Tabelle mit synthetischen Lokalitaetsdaten
fuer Tests, Demos oder Entwicklungen. Die Daten enthalten Namen,
Geometrietypen, Flaechen, Hoehenangaben und Koordinaten (inkl. WKT-Geometrien).
}
\details{
Die Zahl der Lokalitaeten ist auf maximal 20 beschraenkt.
}
