% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_names_field}
\alias{generate_names_field}
\title{Generate Randomized Name Fields}
\usage{
generate_names_field(
  n = 1,
  pool_first = c("Anna", "Ben", "Clara", "David", "Eva", "Felix", "Greta", "Hans", "Ida",
    "Jonas"),
  pool_last = c("Mueller", "Schmidt", "Schneider", "Fischer", "Weber", "Meyer", "Wagner",
    "Becker", "Hoffmann", "Schaefer"),
  min_names = 1,
  max_names = 3,
  abbrev_prob = 0.3,
  separator = "; "
)
}
\arguments{
\item{n}{Anzahl der Felder.}

\item{pool_first}{Vektor moeglicher Vornamen.}

\item{pool_last}{Vektor moeglicher Nachnamen.}

\item{min_names}{Minimale Anzahl an Namen pro Feld.}

\item{max_names}{Maximale Anzahl an Namen pro Feld.}

\item{abbrev_prob}{Wahrscheinlichkeit fuer abgekuerzte Vornamen.}

\item{separator}{Trennzeichen fuer mehrere Namen.}
}
\value{
Ein Character-Vektor mit Namenseintraegen.
}
\description{
Erzeugt realistische Namenseintraege (mit Abkuerzungen) fuer Testdaten.
}
