% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_wkt_linestring_vector}
\alias{generate_wkt_linestring_vector}
\title{Generate WKT LineStrings}
\usage{
generate_wkt_linestring_vector(n)
}
\arguments{
\item{n}{Anzahl der Linien.}
}
\value{
Character-Vektor mit LINESTRING-Geometrien.
}
\description{
Erzeugt zufaellige WKT-Linien (LINESTRING).
}
