% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_wkt_multipolygon_vector}
\alias{generate_wkt_multipolygon_vector}
\title{Generate WKT Multipolygons}
\usage{
generate_wkt_multipolygon_vector(n)
}
\arguments{
\item{n}{Anzahl der Multipolygone.}
}
\value{
Character-Vektor mit MULTIPOLYGON-Geometrien.
}
\description{
Erzeugt zufaellige MULTIPOLYGON-Geometrien mit zwei Teilen.
}
