% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_wkt_point_vector}
\alias{generate_wkt_point_vector}
\title{Generate WKT Point Geometries}
\usage{
generate_wkt_point_vector(n)
}
\arguments{
\item{n}{Anzahl der Punkte.}
}
\value{
Character-Vektor mit WKT-POINT-Strings.
}
\description{
Erzeugt zufaellige WKT-Punkte im Bereich Mitteleuropas.
}
