% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_wkt_polygon}
\alias{generate_wkt_polygon}
\title{Generate a Single WKT Polygon or Multipolygon}
\usage{
generate_wkt_polygon()
}
\value{
Ein einzelner WKT-String.
}
\description{
Gibt zufaellig ein einfaches Polygon oder ein Multipolygon zurueck.
}
