% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_wkt_polygon_vector}
\alias{generate_wkt_polygon_vector}
\title{Generate WKT Polygons}
\usage{
generate_wkt_polygon_vector(n)
}
\arguments{
\item{n}{Anzahl der Polygone.}
}
\value{
Character-Vektor mit POLYGON-Geometrien.
}
\description{
Erzeugt einfache WKT-Polygon-Geometrien.
}
