% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator_image.R
\name{get_indicator_image}
\alias{get_indicator_image}
\title{wrapper around \code{change_pixel_color_HVS()} for the intention of clarity.}
\usage{
get_indicator_image(pixel.array, pixel.idx, target.color, mask_extreme)
}
\arguments{
\item{pixel.array}{image array as loaded via duflor::load_image()}

\item{pixel.idx}{list declaring pixels to which target.color.hsv is applied}

\item{target.color}{color in rgb-format \code{0-255}, or a member of  \code{colors()}}

\item{mask_extreme}{by default, only the \code{hue}- and \code{saturation}-components
are applied to the pixels. Set this argument to \code{TRUE} to also apply \code{value}.
This will increase the contrast of the image drastically, but might result in
less favorable images.)}
}
\value{
\code{pixel.array} with hsv-values of pixels at positions \code{pixel.idx} modified.
}
\description{
wrapper around \code{change_pixel_color_HVS()} for the intention of clarity.
}
