% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{get_javaVM_exceptions}
\alias{get_javaVM_exceptions}
\title{retrieve exceptions of the JVM}
\usage{
get_javaVM_exceptions()
}
\value{
boolean value whether or not the JVM contained exceptions.
}
\description{
check the JVM for any pending exceptions, clear them and return TRUE
if exceptions were found.
}
\keyword{internal}
