% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_unique_list_elements}
\alias{get_unique_list_elements}
\title{compare two lists and return all keys that are not present in both.}
\usage{
get_unique_list_elements(a, b)
}
\arguments{
\item{a}{a list}

\item{b}{a list}
}
\value{
vector of elements not present in either \code{a} or \code{b}.
}
\description{
the keys' contents are not considered. Only the presence or absence of the
key is relevant.
}
\examples{
duflor:::get_unique_list_elements(list(a = 1,b =2,c= 3),list(a = 1,b = 1,d = 3))
}
\keyword{internal}
