% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrors.R
\name{HSVtoRGB}
\alias{HSVtoRGB}
\title{mirror of \code{imager::HSVtoRGB()}}
\usage{
HSVtoRGB(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[imager:imager.colourspaces]{imager::HSVtoRGB}}
  \describe{
    \item{\code{im}}{an image}
  }}
}
\description{
See \link[imager:imager.colourspaces]{imager::HSVtoRGB}
}
