% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_image}
\alias{load_image}
\title{load image as HSV or RGB-array}
\usage{
load_image(
  image.path,
  subset_only = FALSE,
  return_hsv = TRUE,
  crop_left = 0,
  crop_right = 0,
  crop_top = 0,
  crop_bottom = 0
)
}
\arguments{
\item{image.path}{path to image-file}

\item{subset_only}{do you want to load only a subset range of values}

\item{return_hsv}{TRUE by default. Controls whether or not function returns pixel-data in \code{HSV}-colorspace or in \code{RGB}}

\item{crop_left}{number of pixels to crop from the left edge of the image}

\item{crop_right}{see \code{crop_left}}

\item{crop_top}{see \code{crop_left}}

\item{crop_bottom}{see \code{crop_left}}
}
\value{
hsv-formatted pixel.array, unless \code{HSV==false}.
}
\description{
load image as HSV or RGB-array
}
\note{
\code{load_image()} allows loading subsets of images using \code{imager::load.image()}.
}
