% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{objs}
\alias{objs}
\title{wrapper around \code{object.size()}}
\usage{
objs(x)
}
\arguments{
\item{x}{object to retrive memory size of}
}
\value{
size of \code{x} in GB
}
\description{
wrapper around \code{object.size()}
}
\examples{
duflor:::objs(1:1:500000)
}
\keyword{internal}
