% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_array_as_image_sRGB.R
\name{plot_array_as_image_sRGB}
\alias{plot_array_as_image_sRGB}
\title{plot an rgb-array via graphics-package}
\usage{
plot_array_as_image_sRGB(rgb.array, main = "title")
}
\arguments{
\item{rgb.array}{rgb-formatted image-array.}

\item{main}{title to pot above the plot as a sub-title. is usually not rendered properly}
}
\value{
/
}
\description{
plot an rgb-array via graphics-package
}
\examples{
\dontrun{
plot_array_as_image_sRGB(
    HSVtoRGB(
        imager::load.example("parrots")
        ),
    main = str_c("parrots-example")
)
}
}
