% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{prep_loading}
\alias{prep_loading}
\title{Set java-parameters required for VM-setup.}
\usage{
prep_loading(gb = 8)
}
\arguments{
\item{gb}{integer number of gigabytes of heap-space to assign to the Java environment.}
}
\value{
call to \code{check_javaVM_setup()}
}
\description{
Set java-parameters required for VM-setup.
}
