% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{reassign_integers_by_frequency}
\alias{reassign_integers_by_frequency}
\title{takes a cluster-range of integers and reassigns them based on their frequency}
\usage{
reassign_integers_by_frequency(clus)
}
\arguments{
\item{clus}{vector of integers.}
}
\value{
frequency-based reassigned instance of \code{clus}
}
\description{
For a input \code{clus = c(1,1,1,2,2,3,3,3,3)}, returns \code{clus = c(2,2,2,3,3,1,1,1,1)}.
}
\examples{
\dontrun{
clus <- c(1,1,1,2,2,3,3,3,3)
clus_new <- reassign_integers_by_frequency(clus)
print(clus)
print(clus_new)
}
}
\keyword{internal}
