% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rectangularRange_HSV_cpp2}
\alias{rectangularRange_HSV_cpp2}
\title{'C++'-implementation of \code{\link[=rectangularRange_HSV]{rectangularRange_HSV()}}}
\usage{
rectangularRange_HSV_cpp2(H, S, V, upper_bound, lower_bound, image_width)
}
\arguments{
\item{H}{respective component of a \code{pixel.array}}

\item{S}{respective component of a \code{pixel.array}}

\item{V}{respective component of a \code{pixel.array}}

\item{upper_bound}{EITHER:
\itemize{
\item list of upper HSV-bounds, e.g. \code{list(green = c(H_green_lower,S_green_lower,V_green_lower),drought = c(H_drought_lower,S_drought_lower,V_drought_lower))}
\item single vector of length 3 declaring a set of HSV-values
}}

\item{lower_bound}{see \code{upper_bound}}

\item{image_width}{Width of \code{pixel.array}, as returned via \code{dim(pixel.array)[1]}}
}
\value{
A list-object with the following elements (when supplying one one pair of bounds)
\itemize{
\item \code{pixel.idx} - pixel-locations of pixels detected between lower and upper bound.
}

Upon failure to find any matching pixels, an empty matrix of dimensions \verb{[0, 1:2]} is returned.
}
\description{
'C++'-implementation of \code{\link[=rectangularRange_HSV]{rectangularRange_HSV()}}
}
\note{
The use of  \code{\link[=rectangularRange_HSV]{rectangularRange_HSV()}} is strongly discouraged in favour of this function,
due to its drastically slower execution.
Unlike \link{rectangularRange_HSV_cpp}, this function will threshold on the Value-parameter as well.
}
