% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{reinitialise_javaVM}
\alias{reinitialise_javaVM}
\title{Reinitalises JVM with previously used parameters}
\usage{
reinitialise_javaVM(heapspace = NA, unit = NA, pkg)
}
\arguments{
\item{heapspace}{integer heapspace to be allocated to the new JVM. If omitted, the contents of \code{getOption("duflor..used_JVM_heapspace")} is used.}

\item{unit}{unit of \code{heapspace}. Must be either of \code{c("g","m")}. If omitted, the contents of \code{getOption("duflor..used_JVM_heap_unit")} is used.}

\item{pkg}{return-value of \code{packageName()}}
}
\description{
Used internally by \code{\link[=load_image]{load_image()}}.
If \code{\link[=get_javaVM_exceptions]{get_javaVM_exceptions()}} found any exceptions, this function is called
and re-initialises the JVM with the provided values.
issues a warning suggesting to re-initialise the value before loading the package.
}
\keyword{internal}
