% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{replace_umlaute}
\alias{replace_umlaute}
\title{Replace German Umlauts}
\usage{
replace_umlaute(text)
}
\arguments{
\item{text}{Character-Vektor mit deutschen Umlauten.}
}
\value{
Character-Vektor mit ersetzten Umlauten.
}
\description{
Ersetzt deutsche Umlaute in Texten mit ASCII-kompatiblen Umschriften.
}
