% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_adjacency_coords.R
\name{retrieve_adjacency_coords}
\alias{retrieve_adjacency_coords}
\title{return coordinates by cluster_id from \code{pixel.idx}}
\usage{
retrieve_adjacency_coords(pixel.idx, cluster_id)
}
\arguments{
\item{pixel.idx}{list declaring pixels to which target.color.hsv is applied}

\item{cluster_id}{index of the to-be-retrieved cluster}
}
\value{
\code{pixel.idx} with added 3rd column \code{clus} mapping to a cluster
}
\description{
return coordinates by cluster_id from \code{pixel.idx}
}
\examples{
\dontrun{
pixels <- extract_pixels_HSV(...) # extract pixels of a certain color-range
adjacency <- adjacency(pixels$identifier$pixel.idx) # assign clusters
coords <- retrieve_adjacency_coords(adjacency,1) # retrieve coordinates of first cluster

plot_array_as_image_sRGB( # display result
    HSVtoRGB(
        apply_HSV_color_by_mask(
            pixel.array,
            coords,
            target.color = "white",
            mask_extreme = T
            )
        )
    )
}
}
