% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrors.R
\name{RGBtoHSV}
\alias{RGBtoHSV}
\title{mirror of \code{imager::RGBtoHSV()}}
\usage{
RGBtoHSV(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[imager:imager.colourspaces]{imager::RGBtoHSV}}
  \describe{
    \item{\code{im}}{an image}
  }}
}
\description{
See \link[imager:imager.colourspaces]{imager::RGBtoHSV}
}
