% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{RGBtosRGB}
\alias{RGBtosRGB}
\title{mirror of \code{imager::sRGBtoRGB()}}
\usage{
RGBtosRGB(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[imager:imager.colourspaces]{imager::RGBtosRGB}}
  \describe{
    \item{\code{im}}{an image}
  }}
}
\description{
See \link[imager:imager.colourspaces]{imager::sRGBtoRGB}
}
