% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{setup_javaVM}
\alias{setup_javaVM}
\title{Set java-parameters required for VM-setup.}
\usage{
setup_javaVM(gb = 8)
}
\arguments{
\item{gb}{integer number of gigabytes of heapspace to assign to the Java environment.}
}
\value{
call to \code{check_javaVM_setup()}
}
\description{
Used internally by \code{\link[=load_image]{load_image()}}.
If the returned heapspace size used for initialisation is less than 8 gb,
issues a warning suggesting to re-initialise the value before loading the package.
}
