% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duflor-package.R
\docType{package}
\name{duflor-package}
\alias{duflor}
\alias{duflor-package}
\title{duflor: Plant Image Analysis For Determination of Leaf- and Root-Area}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Determine 2-dimensional plant area via known area of an identifier dot.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://Claudius-Appel.github.io/duflor/}
  \item Report bugs at \url{https://github.com/Claudius-Appel/duflor/issues}
}

}
\author{
\strong{Maintainer}: Claudius Appel \email{claudius.appel@freenet.de}

}
\keyword{internal}
