% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duflor-package.R
\docType{package}
\name{dufloR-package}
\alias{dufloR}
\alias{dufloR-package}
\title{dufloR: Ein Paket zur Validierung floristischer Funddaten in R}
\description{
dufloR ist ein R-Paket zur Validierung floristischer Funddaten gemaess dem offiziellen
DUFLOR-Datenstandard (Datenstandard fuer uniforme floristische Funddaten, Bundesamt fuer Naturschutz).
}
\details{
Das Paket bietet Funktionen zur:
- strukturierten Pruefung von Beobachtungs- und Nachweisdaten gegen ein JSON-Schema
- Erstellung synthetischer Testdatensaetze
- Generierung typischer Validierungsfehler zur Testabdeckung
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.deinprojekturl.de/dufloR}
  \item Report bugs at \url{https://gitlab.deinprojekturl.de/dufloR/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Ruhm \email{jonathan.ruhm@bfn.de}

}
\keyword{internal}
