% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pixels_HSV.R
\name{extract_pixels_HSV}
\alias{extract_pixels_HSV}
\title{Extract pixels from image which fall inbetween lower- and upper bounds.}
\usage{
extract_pixels_HSV(
  pixel.array,
  lower_bound,
  upper_bound,
  fast_eval = TRUE,
  bundle_pixelarray = FALSE,
  check_value = FALSE
)
}
\arguments{
\item{pixel.array}{image array as loaded via duflor::load_image()}

\item{lower_bound}{see \code{upper_bound}}

\item{upper_bound}{EITHER:
\itemize{
\item list of upper HSV-bounds, e.g. \code{list(green = c(H_green_lower,S_green_lower,V_green_lower),drought = c(H_drought_lower,S_drought_lower,V_drought_lower))}
\item single vector of length 3 declaring a set of HSV-values
}}

\item{fast_eval}{Use \verb{C++}-code where possible to reduce execution time?}

\item{bundle_pixelarray}{logical, indicating if the input parameter \code{pixel.array} is to be bundled into the return-value
This is useful to retain \code{pixel.array} into the output if this function is called in a loop.}
}
\value{
EITHER:
list-object with the following elements (when supplying one one pair of bounds)
\itemize{
\item \code{pixel.idx} - pixel-locations of pixels detected between lower and upper bound.
\item \code{pixel.count} - number of pixels detected between lower and upper bound
\item \code{img.fraction} - fraction of image detected between lower and upper bound
\item \code{original.img} - fed-in pixel.array
or
list-object with the following elements (when supplying one one pair of bounds)
\item \code{pixel.idx} - pixel-locations of pixels detected between lower and upper bound.
\item \code{pixel.count} - number of pixels detected between lower and upper bound
\item \code{img.fraction} - fraction of image detected between lower and upper bound
\item \code{original.img} - fed-in pixel.array
\item \code{indicator.img} - \code{pixel.array} with hsv-values of pixels at positions \code{pixel.idx} modified.
}
}
\description{
Extract pixels from image which fall inbetween lower- and upper bounds.
}
