% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{norm_to_range_01}
\alias{norm_to_range_01}
\title{normalise a vector to range \code{0-1}}
\usage{
norm_to_range_01(x)
}
\arguments{
\item{x}{vector to normalise to range \code{0-1}}
}
\value{
vector, normalised to range \code{0-1}
}
\description{
normalise a vector to range \code{0-1}
}
\keyword{internal}
